Change testers port to 8080.
