+++ src/components/content/hr/HRSpecific/publicHolidays/publicHolidays.css
@@ -14,10 +14,6 @@
   border-radius: 0.3em;
 } */
 
-.inlineItem{
-  display: inline;
-}
-
 .positionBtn{
   margin-left: 20%;
 }
