+++ staticServer.js
@@ -12,6 +12,6 @@ app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, 'build/index.html'));
 });
 
-app.listen(8081, () => {
+app.listen(8080, () => {
   console.log("Serving static files .... *** ###");
 });
